requirements.txt
django==4.2.3
requests==2.28.0
sqlparse==0.4.2
typing_extensions==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.1.0
django-cors-headers
boto3==1.24.94
django-environ==0.4.5
mysqlclient==2.0.3
django-jet-reboot
uritemplate
pyyaml
drf-spectacular
APScheduler==3.10.4
SQLAlchemy==2.0.22
pymongo[gssapi,srv,tls]
dnspython
fiscalyear